requirements.txt
confluent_kafka
simplejson
psycopg2-binary
requests
pyspark==3.5